Added tests and poetry dependency management
