+++ gqlalchemy/transformations/translators/dgl_translator.py
@@ -1,4 +1,16 @@
-from typing import List
+# Copyright (c) 2016-2022 Memgraph Ltd. [https://memgraph.com]
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 import dgl
 import torch
@@ -102,15 +114,14 @@ class DGLTranslator(Translator):
         for node_label, features_dict in node_features.items():
             for feature_name, features in features_dict.items():
                 features = Translator.validate_features(features, graph.num_nodes(node_label))
-                if not features is None:
+                if features is not None:
                     graph.nodes[node_label].data[feature_name] = features
 
         # Set edge features
         for edge_triplet, features_dict in edge_features.items():
             for feature_name, features in features_dict.items():
                 features = Translator.validate_features(features, graph.num_edges(edge_triplet))
-                if not features is None:
+                if features is not None:
                     graph.edges[edge_triplet].data[feature_name] = features
 
         return graph
-
