+++ pyproject.toml
@@ -32,7 +32,7 @@ exclude = '''
 '''
 
 [tool.poetry.dependencies]
-python = "^3.7"
+python = "^3.8"
 pymgclient = "1.3.1"
 networkx = "^2.5.1"
 pydantic = "^1.8.2"
@@ -42,6 +42,10 @@ dacite = "^1.6.0"
 adlfs = "^2022.2.0"
 neo4j = "^4.4.3"
 docker = "^5.0.3"
+torch = "^1.13.1"
+numpy = "^1.24.1"
+dgl = "^0.9.1"
+poethepoet = "^0.18.1"
 
 [tool.poetry.group.dev.dependencies]
 black = "^22.3.0"
@@ -51,6 +55,9 @@ pytest-cov = "^2.12.0"
 pytest-flake8 = "1.0.7"
 pytest-timeout = "^1.4.2"
 pre-commit = "^2.15.0"
+
+[tool.poe.tasks]
+install-pyg-cpu = "pip install torch-scatter torch-sparse torch-cluster torch-spline-conv torch-geometric"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
