+++ tests/transformations/test_pyg_transformations.py
@@ -1,3 +1,17 @@
+# Copyright (c) 2016-2022 Memgraph Ltd. [https://memgraph.com]
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 from typing import Dict, Any, Set
 from numbers import Number
 
@@ -18,6 +32,7 @@ from tests.transformations.utils import init_database
 ##########
 # UTILS
 ##########
+
 
 def _check_entity_exists_in_pyg(
     entity_data, entity_mark: str, properties: Dict[str, Any], translated_properties: Set[str] = {}
